Add new Image Processing stuff

+++ src/J-Core/Gui/IGuiExtras.cpp
@@ -491,11 +491,11 @@ namespace JCore::Gui {
         window->DrawList->PathClear();
 
         float time = float(g.Time);
-        int num_segments = 30;
-        int start = abs(ImSin(time * 1.8f) * (num_segments - 5));
+        float num_segments = 30;
+        float start = abs(ImSin(time * 1.8f) * (num_segments - 5));
 
-        const float a_min = IM_PI * 2.0f * ((float)start) / (float)num_segments;
-        const float a_max = IM_PI * 2.0f * ((float)num_segments - 3) / (float)num_segments;
+        const float a_min = IM_PI * 2.0f * (start) / num_segments;
+        const float a_max = IM_PI * 2.0f * (num_segments - 3) / num_segments;
 
         const ImVec2 centre = ImVec2(pos.x + radius, pos.y + radius + style.FramePadding.y);
